# packages/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Nexus API
  description: Local-first AI document analysis and knowledge graph platform
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: local development environment
paths:
...
  /workspaces/{workspaceId}/chats:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Workspace ID

    get:
      summary: List all chats in workspace
      operationId: listChats
      tags: [chats]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [chats, total]
                properties:
                  chats:
                    type: array
                    items:
                      $ref: '#/components/schemas/Chat'
                  total:
                    type: integer
                    description: Total number of chats
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      summary: Create a new chat
      operationId: createChat
      tags: [chats]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                  description: Chat title (user-defined or auto-generated)
                filter_config:
                  $ref: '#/components/schemas/FilterConfig'
      responses:
        '201':
          description: Chat created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/chats/{chatId}:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: chatId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get chat details with message history
      operationId: getChat
      tags: [chats]
      parameters:
        - name: message_limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          description: Number of recent messages to return
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [chat, messages]
                properties:
                  chat:
                    $ref: '#/components/schemas/Chat'
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a chat (soft delete)
      operationId: deleteChat
      tags: [chats]
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/chats/{chatId}/messages:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: chatId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      summary: Send a message and get AI response
      operationId: sendMessage
      tags: [chats]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
                  description: User message content
      responses:
        '201':
          description: Message sent and response generated
          content:
            application/json:
              schema:
                type: object
                required: [user_message, assistant_message]
                properties:
                  user_message:
                    $ref: '#/components/schemas/ChatMessage'
                  assistant_message:
                    $ref: '#/components/schemas/ChatMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ========================================
  # Analysis Endpoints
  # ========================================
...
components:
  schemas:
...
    Chat:
      type: object
      required: [id, workspace_id, title, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        title:
          type: string
        filter_config:
          $ref: '#/components/schemas/FilterConfig'
        message_count:
          type: integer
          minimum: 0
          description: Total number of messages in this chat
        last_message_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the last message
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChatMessage:
      type: object
      required: [id, chat_id, role, content, message_index, created_at]
      properties:
        id:
          type: string
          format: uuid
        chat_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        message_index:
          type: integer
          minimum: 0
          description: Sequential index within the chat (0-based)
        document_refs:
          type: array
          items:
            $ref: '#/components/schemas/DocumentReference'
          description: References to source documents (RAG citations)
        created_at:
          type: string
          format: date-time
    DocumentReference:
      type: object
      required: [document_id, chunk_index, score]
      properties:
        document_id:
          type: string
          format: uuid
        document_name:
          type: string
        chunk_index:
          type: integer
          minimum: 0
        # ページ番号を追加：フロントのソースパネルで「P.3」表示に使う
        page_number:
          type: integer
          minimum: 1
          nullable: true
          description: Page number in the source document (1-based)
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Relevance score (0.0 - 1.0)
        # content_preview は廃止せず残すが、上限を緩和しない
        # 全文表示は GET /documents/{id}/pages/{page} で別途取得する設計（関心の分離）
        content_preview:
          type: string
          maxLength: 200
          description: Short preview of the chunk content
...
    FilterConfig:
      type: object
      properties:
        document_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Specific documents to search (empty = all)
        directory_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Search within specific directories
        tags:
          type: array
          items:
            type: string
          description: Filter by document tags
      description: Configuration for filtering RAG search scope

    # ========================================
    # Analysis Schemas
    # ========================================
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: 
              type: string
            message: 
              type: string
            details: 
              type: object  
            timestamp: 
              type: string
            request_id: 
              type: string  
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'